requirements.txt
flask
tensorflow
Pillow
numpy
opencv-python
werkzeug
gdown
gunicorn